refactor(parties,invoicing): move query-key factories to hooks/query-keys.ts
buildPartiesQueryKey and buildInvoicingQueryKey lived in providers/;
relocate them to the canonical hooks/query-keys.ts (matching taxonomy,
workflow, notifications). Public export surface is unchanged.
