+++ packages/@granit/react-parties/src/__tests__/parties-provider.test.tsx
@@ -3,11 +3,8 @@ import * as React from 'react';
 import { describe, expect, it } from 'vitest';
 
 import { DEFAULT_BASE_PATH } from '../constants';
-import {
-  PartiesProvider,
-  buildPartiesQueryKey,
-  usePartiesConfig,
-} from '../providers/parties-provider';
+import { buildPartiesQueryKey } from '../hooks/query-keys';
+import { PartiesProvider, usePartiesConfig } from '../providers/parties-provider';
 
 import type { PartiesConfig } from '../providers/parties-provider';
 import type { AxiosInstance } from 'axios';
