+++ packages/@granit/react-parties/src/hooks/use-parties.ts
@@ -25,7 +25,9 @@ import {
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
 import { logger } from '../logger';
-import { buildPartiesQueryKey, usePartiesConfig } from '../providers/parties-provider';
+import { usePartiesConfig } from '../providers/parties-provider';
+
+import { buildPartiesQueryKey } from './query-keys';
 
 import type {
   CreatePartyOptions,
