+++ packages/@granit/react-parties/src/providers/parties-provider.tsx
@@ -33,12 +33,3 @@ export const PartiesProvider = Provider;
 
 /** Returns the parties configuration from the nearest `PartiesProvider`. */
 export const usePartiesConfig = useConfig;
-
-/** Builds a consistent React Query key for parties operations. */
-export function buildPartiesQueryKey(
-  config: PartiesConfig,
-  ...segments: readonly unknown[]
-): readonly unknown[] {
-  const prefix = config.queryKeyPrefix ?? ['parties'];
-  return [...prefix, ...segments];
-}
